Suboptimal allocation of BloomFilter bitarray size
Empirical measurement of the false positive rate of BloomFilter is
substantially lower that the requested error rates. For example, a
BloomFilter with capacity 20,000 and an error rate of .01, gives actual
error rate .0003. This is because the bit size allocated to the filter
is unnecessarily large, and does not match the theoretically calculated
required size.